Got Dispatcher working with URL requests
Now MainApp calls the dispatcher, passes it the request.
The Dispatcher breaks the request up into "parts", and
checks the first part to see if it's a class.

If it is, we'll assume it's a controller class,
(later it might be worth checking!) and we'll pass control
to that.

Other changes: Added a timer and RAM usage to check I'm not
doing anything too crazy like...