Got basic models working, including test examples
I've got a connection to the database (good!)
The connection is run only when needed and only once, thanks
to the static methods in DbControl.  Did some very basic testing
and this method seems pretty efficient.

I've set up a template Model, and derived two models from it:
TodoItem and User.  Both of these have corresponding database
tables.

In MainApp I have the basic scripts there to test these out
for basic SPOD actions like Create, Read, List, Search.
(It's all fine, I'll assume Update and Delete are rocking too).

I've tested and MySQL is working with the foreign keys (ie It
won't let me insert with a bad userID, and if I delete or modify
a User, then it cascades to the relevant rows in TodoItem)

Next step - see if I can set these relations up in Haxe so the
objects have them, and note the effect on performance.