Tested hasMany relationships.
They all worked.  Sample code is in the documentation for
BaseDbModel.  That documentation has been expanded a bit.

It might be worth seeing if we can get a hasMany thing
built into Haxe SPOD.  We'll see though.

Working well, performance is fine (I think - not my specialty)

Haven't figured out a many-to-many relationship, but I think
it could easily be solved with a join table and a join model.