Added BaseController.hx, which all Controllers should extend from.
Notes so far:
-I was originally going to have controllers static, but it seems
like it'll be harder to check things at runtime, so I may make the
controller a class that you create an object from.
-This has the added advantage of letting us call multiple controllers
and even multiple copies of the same controller in one page load,
without having to worry about conflicts (this could be used for
displaying, for example, menus and slideshows and other controls
not related to our page request, but that should be used on each
page).
-I think the new() method of each controller should take the request
and delegate it to the appropriate action(), passing on all the
parameters as it goes.
-The appropriate action will talk to the load the appropriate view,
and add all the template variables etc.
-The controller object should have a toString() method that we can
easily use, rather than outputting everything directly to the browser.
This will allow us to do caching and other such things on an application
level, not a controller level.