Added HxTpl methods to let Controller control the View output.
Each action can call "loadTemplate()".
If a url is given, we'll load that.  If not, we'll try load
a template according to convention.

When done, you call "printTemplate()" and it saves that to
our output string, which is printed by the Dispatcher.

This way if a controller wants to ignore our templating
engine, they can.  But it should be very easy (two lines
that don't even need to change) to get the templating happening
for each action.

I still need to test this works.  Then - start making a real app.