Now to support Site-wide and controller-wide templates
The BaseController now doesn't just print the template for the
requested action, it searches for an appropriate template for
the rest of the page.  This might be defined for the whole
Controller (this option takes precedence), the whole App, or not
at all, in which case we just fall back to printing the View only.